# v1/auth/openapi.yaml
openapi: 3.1.2
info:
  title: Auth
  description: Authentication API for aces.hackclub.com
  version: 0.1.0

# servers:
#   - url: http://api.example.com/v1
#     description: Optional server description, e.g. Main (production) server
#   - url: http://staging-api.example.com
#     description: Optional server description, e.g. Internal staging server for testing

paths:
  /auth/refresh_session:
    post:
      summary: Refreshes the expiration date for the users JWT token
      description: Requires a preexisting JWT
      security:
        - bearerAuth: []
      responses:
        "200": # status code
          description: Successfully refreshed JWT token
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: If the operation succeeded
                    example: true
                  # token_type:
                  #   type: string
                  #   example: "bearer"
                  # expires_in:
                  #   type: integer
                  #   description: Token expiration time in seconds
                  #   example: 3600
        "401":
          description: The provided JWT was invalid
        "400":
          description: Request body did not match expected body

  /auth/send_otp:
    post:
      summary: Sends an OTP to the provided 
      requestBody:
        description: The email to send the OTP to
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: 
                email:
                  type: string
                  example: "example@example.com"
      responses:
        "200":
          description: Successfully sent the OTP
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: If the operation succeeded
                    example: true
        "400":
          description: Request body did not match expected body

  /auth/validate_otp:
    post:
      summary: Validates a sent OTP
      requestBody:
        description: What email and OTP to cehck for
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: example@example.com
                otp:
                  type: string
                  example: "012345"
      responses:
        "200":
          description: OTP Matched and JWT was issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    description: If the operation succeeded
                    example: true
        "401":
          description: The OTP did not match the provided email
        "500":
          description: An error occured while checking
            # schema:
            #   $ref: "#/components/schemas/otpRequest"
        "400":
          description: Request body did not match expected body

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT